FP plumbing: GPR<->FPR moves + FP loads/stores (FR=1), checker fixed for mips3/4
Graft the FP move/load/store datapath from mipscore@superscalar into r9999.
Actual FPU arithmetic is deferred; this is moves + ld/st only, FR=1 flat
32x64-bit FP registers (mips3/4), validated by directed co-sim tests.

RTL (rename + mem pipe + L1D):
- uop.vh/rob.vh/core.sv: FP register file + rename (FP free list, FP RAT,
  per-source int-vs-FP mux, ROB valid_fp_dst, retire-free). MEM_MOV op;
  fp_dst threaded through mem_req_t/mem_rsp_t.
- exec.sv: FP PRF (r_fp_prf) + inflight; GPR<->FPR moves routed through the
  L1D as MEM_MOV carrying data in the request addr field (mtc1/mfc1 =
  sign_extend32(src[31:0])); FP loads (lwc1/ldc1) write the FP PRF; FP store
  data sourced from the FP PRF via the store-data queue (dq_t.fp).
- l1d.sv: MEM_MOV intercepted before TLB/cache (echo addr -> dst PRF with
  fp_dst select); port2 load response routes fp_dst.
- decode_mips.sv: mtc1/mfc1 (flat FR=1), lwc1/swc1, ldc1/sdc1 (64b-mode-gated).

Checker (interpret.cc/.hh): cpr1 was uint32_t[32] (FR=0 even/odd pairs = the
mips32 model) while the RTL is FR=1. Converted cpr1 to uint64_t[32]; mtc1/mfc1
now sign-extend bits[31:0] (mfc1 previously zero-extended); ldc1/sdc1 address a
single 64-bit register (not an even/odd pair); removed FR=0 leftovers from the
arith/cvt helpers (untested - no RTL FPU yet).

Tests (tests/fpu/, co-sim clean with -c 1): fp_move, fp_load, fp_store, fp_sext.
33/33 integer randgen rt_* co-sim still pass.

+++ interpret.hh
@@ -110,7 +110,9 @@ public:
   reg_t lo = 0;
   reg_t hi = 0;
   uint32_t cpr0[32] = {0};
-  uint32_t cpr1[32] = {0};
+  /* FR=1 (mips3/mips4): 32 independent 64-bit FP registers, NOT FR=0 even/odd
+   * 32-bit pairs.  Singles live in bits[31:0]; doubles use the full 64 bits. */
+  uint64_t cpr1[32] = {0};
   uint32_t fcr1[5] = {0};
   uint64_t icnt = 0;
   uint8_t brk = 0;
